updated contracts da sa vedem

+++ contracts/lib/UniswapV2Library.sol
@@ -23,13 +23,15 @@ library UniswapV2Library {
     ) internal pure returns (address pair) {
         (address token0, address token1) = sortTokens(tokenA, tokenB);
         pair = address(
-            uint256(
-                keccak256(
-                    abi.encodePacked(
-                        hex"ff",
-                        factory,
-                        keccak256(abi.encodePacked(token0, token1)),
-                        hex"96e8ac4277198ff8b6f785478aa9a39f403cb768dd02cbee326c3e7da348845f" // init code hash
+                    uint160(
+                        uint256(
+                            keccak256(
+                                abi.encodePacked(
+                                    hex"ff",
+                                    factory,
+                                    keccak256(abi.encodePacked(token0, token1)),
+                                    hex"96e8ac4277198ff8b6f785478aa9a39f403cb768dd02cbee326c3e7da348845f" // init code hash
+                        )
                     )
                 )
             )
